Removes User Profile Landing
